v2

+++ src/main/java/lk/probidder/service/impl/UserServiceImpl.java
@@ -5,6 +5,7 @@ import lk.probidder.dto.UserDTO;
 import lk.probidder.entity.User;
 import lk.probidder.repo.UserRepo;
 import lk.probidder.service.UserService;
+import lk.probidder.service.exception.DuplicateException;
 import lk.probidder.util.mapper.UserMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -51,12 +52,20 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public String createUser(UserDTO urd) {
+    public String createUser(UserDTO urd) throws DuplicateException {
+        Optional<User> byId = userRepository.findById(urd.getId());
+        if (byId.isPresent()) {
+            throw new DuplicateException();
+        }
         return userRepository.save(userMapper.toUser(urd)).getId() + " saved";
     }
 
     @Override
-    public void deleteUser(Long id) {
+    public void deleteUser(Long id) throws ClassNotFoundException {
+        Optional<User> byId = userRepository.findById(id);
+        if (byId.isEmpty()) {
+            throw new ClassNotFoundException();
+        }
         userRepository.deleteById(id);
     }
 }
